requirements.txt
fastapi
uvicorn
requests
mistralai
pandas
numpy
rapidfuzz
praw
python-dotenv
cloudscraper
beautifulsoup4
psycopg2-binary
sqlalchemy